+++ dropwizard-guicey/src/main/java/ru/vyarus/dropwizard/guice/test/jupiter/ext/track/TrackersSupport.java
@@ -72,6 +72,7 @@ public class TrackersSupport extends AnnotatedTestFieldSetup<TrackBean, Tracker>
     }
 
     @Override
+    @SuppressWarnings("unchecked")
     protected void validateBinding(final EventContext context, final AnnotatedField<TrackBean, Tracker> field) {
         final Tracker tracker = Preconditions.checkNotNull(field.getCustomData(FIELD_TRACKER));
         final Binding binding = context.getInjector().getBinding(tracker.getType());